Optimize locking
The previous locking strategy required taking a mutex regardless of
whether the record is accepted by the filter. This leads to a lot of
synchronization in multithreaded programs making extensive use of log
macros. Since a common code path is calling log, applying a filter, and
throwing the record away, it makes sense to be able to do this in
parallel. To accomplish this, the `inner` Mutex was changed to a RwLock.
Since `&mut self` methods can't be called with a read lock, the Append,
Filter, and Encode traits need to implement Sync, and their implementors
must use interior mutability where needed.