WIP: Adding ability to configure colors for console
Specifically this useful for dark console backgrounds where the default
for 'Blue' for log level INFO is difficult to read.

Most of the changes are in `src/encode/patter/mod.rs`. I added default
parameters for deserializing. I added a function to describe the default
pattern string, and defined that at the top of the source file.
I added a `color_map` to the Encoder structs which is a HashMap that
defines the LogLovel->Color mapping. The color_map is of type
`HashMap<Level, Option<Color>>`, if the Option is None, then no styling
is applied.

TODO: Adding a feature means adding documentation, which isn't complete.
I can do this and add to the PR, however before i continue i want to
make sure this is the direction that the maintiner wants to take.

TODO: Other testing may need to be done. I was testing with `cargo test
--features file` with a modified config option in the yaml test config.
As i am not familiar with this code base there may be other cases that
need to be tested.

TODO/Future: This patch only allows foreground color to be specified.
There are other aspects of styling like bold,italics,background color.
Does it make sense to allow for more detailed styling? That involves
more work and more work for the maintiner. I think forground color will
probably cover most everyone's cases.