Update expand_env_vars to use Into<String>
This change updates expand_env_vars to accept any type implementing
Into<String> instead of just PathBufs. Functionally theres really no change
as the first thing the function did was convert the pathbuf into a
string.

This makes it easier to chain edits to the filename string as the
changes can be made on the string and converted to a Path when we open
the file.