Update expand_env_vars to use Into<Cow>
This change updates expand_env_vars to accept any type implementing
Into<Cow<'str,str> instead of just PathBufs. Functionally theres really
no change as the first thing the function did was convert the pathbuf
into a string which may or may not be mutated. Updating the type to be a
Cow (Clone On Write) smart pointer avoids any allocations for patterns
without environmental mixins.

This makes it easier to chain edits to the filename as the conditional
changes can be made on the Cow and converted to a Path when we open the
file.