Fixed window roller expand env var (#287)
* Add failing test for env var expansion

This test is has a `$ENV{var}` section as part of the pattern string,
right now the FixedWindowRoller does not expand that variable to its
corresponding env variable. This will make the test fail because the
expected rolled file path and actual rolled file path are different.

* Update expand_env_vars to use Into<Cow>

This change updates expand_env_vars to accept any type implementing
Into<Cow<'str,str> instead of just PathBufs. Functionally theres really
no change as the first thing the function did was convert the pathbuf
into a string which may or may not be mutated. Updating the type to be a
Cow (Clone On Write) smart pointer avoids any allocations for patterns
without environmental mixins.

This makes it easier to chain edits to the filename as the conditional
changes can be made on the Cow and converted to a Path when we open the
file.

* Update rotate to take into account env vars

This commit updates the rotate function to expand any environmental
variable sections in the pattern argument.